separate unit tests and integration tests (#184)
